More Expression Parser Work
The expression evaluation code has been switched to  use integers instead of floats.

Since expressions are used only for referencing memory addresses (and those are integers) floats currently have no use. Removing them simplifies the code.

Added basic tests for the expression parser.

+++ VM/Expressions/Node.cs
@@ -4,6 +4,6 @@ namespace VMCore.Expressions
 {
     public abstract class Node
     {
-        public abstract float Evaluate(CPU cpu);
+        public abstract int Evaluate(CPU cpu);
     }
 }
