+++ VM/Expressions/NodeUnary.cs
@@ -14,7 +14,7 @@ namespace VMCore.Expressions
             _op = op;
         }
 
-        public override float Evaluate(CPU cpu)
+        public override int Evaluate(CPU cpu)
         {
             var rhsVal = _rhs.Evaluate(cpu);
